requirements.txt
streamlit
pymongo
pypdfium2
easyocr
pdfplumber
Pillow
docx2txt
pandas
openpyxl
torch  
